requirements.txt
Flask>=2.2
Flask-Limiter>=3.4
openai>=1.0
python-dotenv>=1.0 